requirements.txt
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
idna==2.8
Pillow==6.2.1
requests==2.22.0
soupsieve==1.9.5
urllib3==1.25.7
